+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## v0.3.0 (2020-??-??)
 
+* Added support for multiple-answers questions.
 * Added support for comments at the top level of quiz files (outside Markdown
   content like questions, choices, or feedback).  HTML comments within
   Markdown are now stripped and no longer appear in the final QTI file (#2).